fix: 워크스페이스 페이지 styles.ts 분리, 패키지 선언 중 오류 해결
