+++ types/styled-popover-component/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'styled-popover-component';
